Updated Level Loader, fixed falling through floor
Scence_0_Nathan has new code on both the dummy, player and level loader
that seems to have fixed all of the wall collision issues! Dummy code in
other scenes is messed up, but it doesn't really have to work anyways
because we wont have a dummy in the final version.